+++ client/hud/Inventory.cpp
@@ -164,7 +164,7 @@ void Inventory::updateReals(sf::RenderWindow &window, sf::Time &gameTime)
     float naviconZ = Hud::getDrawDepth(Hud::zindex_frames);
     bit::Vertex3* naviconQuad = &hud->interfaceVertexMap.vertexArray[naviconQuadIndex];
     naviconSprite->applyToQuad(naviconQuad);
-    bit::VertexHelper::positionQuad(naviconQuad, left, top - naviconSprite->height * elementScale, naviconZ, naviconSprite->width, naviconSprite->height, elementScale);
+    bit::VertexHelper::positionQuad(naviconQuad, (int)left, (int)(top - naviconSprite->height * elementScale), naviconZ, naviconSprite->width, naviconSprite->height, elementScale);
 
     // Silohuette for Equipment
     int xOff = 40 * elementScale;
@@ -172,7 +172,7 @@ void Inventory::updateReals(sf::RenderWindow &window, sf::Time &gameTime)
     float siloZ = Hud::getDrawDepth(Hud::zindex_frames);
     bit::Vertex3* siloQuad = &hud->interfaceVertexMap.vertexArray[siloQuadIndex];
     siloSprite->applyToQuad(siloQuad);
-    bit::VertexHelper::positionQuad(siloQuad, equipmentPanel->left + xOff, equipmentPanel->top + yOff, siloZ, siloSprite->width, siloSprite->height, elementScale * 2);
+    bit::VertexHelper::positionQuad(siloQuad, (int)(equipmentPanel->left + xOff), (int)(equipmentPanel->top + yOff), siloZ, siloSprite->width, siloSprite->height, elementScale * 2);
 }
