+++ client/hud/InventoryItemLabel.cpp
@@ -291,7 +291,7 @@ void InventoryItemLabel::updateReals(sf::RenderWindow &window, sf::Time &gameTim
     bit::Element::updateReals(window, gameTime);
 
     float z = this->draggable->isDragging ? Hud::getDrawDepth(Hud::zindex_iconsDragged) : Hud::getDrawDepth(Hud::zindex_icons);
-    icon->position(left, top, z, 64, 64, elementScale);
+    icon->position((int)left, (int)top, z, 64, 64, elementScale);
 }
